refactor: Update integer types in models and services for consistency
- Changed ID fields from int64 to int across various models including BaseModel, Device, and Order-related structures to ensure uniformity.
- Updated service methods and DTOs to reflect the new integer type for IDs, enhancing type consistency throughout the codebase.
- Adjusted related logic in handlers and services to accommodate the type changes, ensuring smooth functionality and compatibility.

+++ models/portal/base_model.go
@@ -5,7 +5,7 @@ package portal
 
 // BaseModel 基础模型.
 type BaseModel struct {
-	ID        int64    `gorm:"primaryKey;autoIncrement"`        // 主键ID
+	ID        int      `gorm:"primaryKey;autoIncrement"`        // 主键ID
 	CreatedAt NavyTime `gorm:"column:created_at;type:datetime"` // 创建时间
 	UpdatedAt NavyTime `gorm:"column:updated_at;type:datetime"` // 更新时间
 }
