+++ models/portal/elastic_scaling_order.go
@@ -3,8 +3,8 @@ package portal
 // OrderDevice 订单设备关联表
 type OrderDevice struct {
 	BaseModel
-	OrderID  int64  `gorm:"primaryKey;column:order_id"`
-	DeviceID int64  `gorm:"primaryKey;column:device_id"`
+	OrderID  int  `gorm:"primaryKey;column:order_id"`
+	DeviceID int  `gorm:"primaryKey;column:device_id"`
 	Status   string `gorm:"type:varchar(50);default:'pending'"`
 }
 
@@ -16,14 +16,14 @@ func (OrderDevice) TableName() string {
 // StrategyExecutionHistory 策略执行历史表
 type StrategyExecutionHistory struct {
 	BaseModel
-	StrategyID     int64    `gorm:"column:strategy_id;type:bigint"`           // 策略ID
-	ClusterID      int64    `gorm:"column:cluster_id;type:bigint"`            // 集群ID
+	StrategyID     int    `gorm:"column:strategy_id;type:bigint"`           // 策略ID
+	ClusterID      int    `gorm:"column:cluster_id;type:bigint"`            // 集群ID
 	ResourceType   string   `gorm:"column:resource_type;type:varchar(100)"`   // 资源池名称
 	ExecutionTime  NavyTime `gorm:"column:execution_time;type:datetime"`      // 执行时间
 	TriggeredValue string   `gorm:"column:triggered_value;type:varchar(255)"` // 触发策略时的具体指标值
 	ThresholdValue string   `gorm:"column:threshold_value;type:varchar(255)"` // 触发策略时的阈值设定
 	Result         string   `gorm:"column:result;type:varchar(50)"`           // 执行结果
-	OrderID        *int64   `gorm:"column:order_id;type:bigint"`              // 关联订单ID
+	OrderID        *int   `gorm:"column:order_id;type:bigint"`              // 关联订单ID
 	Reason         string   `gorm:"column:reason;type:text"`                  // 执行结果的原因
 }
 
@@ -35,8 +35,8 @@ func (StrategyExecutionHistory) TableName() string {
 // NotificationLog 通知日志表
 type NotificationLog struct {
 	BaseModel
-	OrderID          *int64   `gorm:"column:order_id;type:bigint"`               // 关联订单ID(可选)
-	StrategyID       *int64   `gorm:"column:strategy_id;type:bigint"`            // 关联策略ID(可选)
+	OrderID          *int   `gorm:"column:order_id;type:bigint"`               // 关联订单ID(可选)
+	StrategyID       *int   `gorm:"column:strategy_id;type:bigint"`            // 关联策略ID(可选)
 	NotificationType string   `gorm:"column:notification_type;type:varchar(50)"` // 通知类型
 	Recipient        string   `gorm:"column:recipient;type:varchar(255)"`        // 接收人信息
 	Content          string   `gorm:"column:content;type:text"`                  // 通知内容
